cluster-available api added (#297)
* cluster-available api

* fixes
