+++ backend/its/manual/handlers/command.go
@@ -1,4 +1,4 @@
-package api
+package handlers
 
 import (
 	"fmt"
